refactor: remove obsolete code paths
